app/api/v1/learn_words/requirements.txt
fastapi==0.99.0
mangum==0.11.0
pydantic==1.10.9
python-dotenv==1.0.0
starlette==0.27.0
typing-extensions==4.9.0
boto3==1.34.34
uvicorn
python-jose[cryptography]==3.3.0
requests==2.31.0 